+++ Backend/src/email_service.js
@@ -1,4 +1,3 @@
-// valeu indiano <3 te amo
 require('dotenv').config();
 const nodemailer = require('nodemailer');
 const { google } = require('googleapis');
